fix: 총 결제금액에 대한 계산 Orders 클래스로 변경
