+++ src/test/java/christmas/domain/EventManagerTest.java
@@ -7,7 +7,6 @@ import static christmas.global.Event.WEEKDAY;
 import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
 
 import christmas.global.Badge;
-import java.util.List;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
@@ -37,11 +36,4 @@ class EventManagerTest {
         }
     }
 
-    @Test
-    @DisplayName("총 결제 금액이 음수이면 0 반환")
-    void calculatePaymentAmount() {
-        int paymentAmount = EventManager.calculatePaymentAmount(new Orders(List.of(new Order("아이스크림", 1))), 10000,
-                20000);
-        assertThat(paymentAmount).isEqualTo(0);
-    }
 }
